requirements.txt
flask>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
scikit-learn>=1.0.0
pandas>=1.5.0
numpy>=1.20.0
requests>=2.28.0
pyyaml>=6.0
torchvision>=0.15.0


